// src/enums/notification.ts
export enum PROVIDER_NOTIFICATION {
    BOOKING_REQUEST = 'New Booking Request',
    BOOKING_CONFIRMATION = 'Booking Confirmation',
    SERVICE_COMPLETED = 'Service Completed',
    URGENT_BOOKING = 'Urgent Booking Alert',
    CONFIRMATION_NEED = 'Provider Confirmation Needed',
}

export enum EMPLOYER_NOTIFICATION {
    BOOKING_CONFIRM = 'Booking Confirmed',
    SERVICE_ACCEPT = 'Service Request Accepted',
    SERVICE_REJECT = 'Service Request Rejected',
    BOOKING_CANCEL = 'Booking Cancellation',
    SERVICE_COMPLETE = 'Service Completed',
}